+++ creme/linear_model/__init__.py
@@ -1,17 +1,29 @@
 """Linear models."""
+from .ffm import FFMClassifier
+from .ffm import FFMRegressor
+from .fm import FMClassifier
 from .fm import FMRegressor
 from .glm import LinearRegression
 from .glm import LogisticRegression
+from .glm import PoissonRegression
+from .hofm import HOFMClassifier
+from .hofm import HOFMRegressor
 from .pa import PAClassifier
 from .pa import PARegressor
 from .softmax import SoftmaxRegression
 
 
 __all__ = [
+    'FFMClassifier',
+    'FFMRegressor',
+    'FMClassifier',
     'FMRegressor',
+    'HOFMClassifier',
+    'HOFMRegressor',
     'LinearRegression',
     'LogisticRegression',
     'PAClassifier',
     'PARegressor',
+    'PoissonRegression',
     'SoftmaxRegression'
 ]