+++ creme/model_selection/sh.py
@@ -96,7 +96,10 @@ def successive_halving(model, param_grid, X_y, metric, budget, eta=2, verbose=Tr
             >>> best_model = model._set_params(**best_params)
             >>> best_model
             Pipeline (
-              StandardScaler (),
+              StandardScaler (
+                with_mean=True
+                with_std=True
+              ),
               LinearRegression (
                 optimizer=Adam (
                   lr=Constant (